refactor: move DatabaseContext to config/databases module
