+++ cli/nao_core/commands/sync/providers/databases/context.py
@@ -1,5 +1,5 @@
 """Re-export DatabaseContext from its canonical location."""
 
-from nao_core.database_context import DatabaseContext
+from nao_core.config.databases.context import DatabaseContext
 
 __all__ = ["DatabaseContext"]
