+++ cli/nao_core/config/databases/bigquery.py
@@ -6,10 +6,10 @@ import ibis
 from ibis import BaseBackend
 from pydantic import Field, field_validator
 
-from nao_core.database_context import DatabaseContext
 from nao_core.ui import ask_select, ask_text
 
 from .base import DatabaseConfig
+from .context import DatabaseContext
 
 logger = logging.getLogger(__name__)
