+++ cli/nao_core/config/databases/databricks.py
@@ -7,10 +7,10 @@ import ibis
 from ibis import BaseBackend
 from pydantic import Field
 
-from nao_core.database_context import DatabaseContext
 from nao_core.ui import ask_text
 
 from .base import DatabaseConfig
+from .context import DatabaseContext
 
 logger = logging.getLogger(__name__)
