+++ cli/nao_core/config/databases/postgres.py
@@ -5,10 +5,10 @@ from ibis import BaseBackend
 from pydantic import Field
 
 from nao_core.config.exceptions import InitError
-from nao_core.database_context import DatabaseContext
 from nao_core.ui import ask_text
 
 from .base import DatabaseConfig
+from .context import DatabaseContext
 
 
 class PostgresDatabaseContext(DatabaseContext):
